cleanup Clippy errors

+++ coincube-gui/src/app/view/active/transactions.rs
@@ -97,7 +97,7 @@ fn transaction_row<'a>(
     let is_receive = matches!(payment.payment_type, PaymentType::Receive);
 
     // Format timestamp
-    let time_text = format_time_ago(payment.timestamp as u32);
+    let time_text = format_time_ago(payment.timestamp);
 
     // Extract description from payment details
     let description = match &payment.details {
